+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/db/PinnedSiteEntity.kt
@@ -7,6 +7,9 @@ package mozilla.components.feature.top.sites.db
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import mozilla.components.feature.top.sites.TopSite
+import mozilla.components.feature.top.sites.TopSite.Type.DEFAULT
+import mozilla.components.feature.top.sites.TopSite.Type.PINNED
 
 /**
  * Internal entity representing a pinned site.
@@ -28,4 +31,25 @@ internal data class PinnedSiteEntity(
 
     @ColumnInfo(name = "created_at")
     var createdAt: Long = System.currentTimeMillis()
-)
+) {
+    internal fun toTopSite(): TopSite {
+        val type = if (isDefault) DEFAULT else PINNED
+        return TopSite(
+            id,
+            title,
+            url,
+            createdAt,
+            type
+        )
+    }
+}
+
+internal fun TopSite.toPinnedSite(): PinnedSiteEntity {
+    return PinnedSiteEntity(
+        id = id,
+        title = title ?: "",
+        url = url,
+        isDefault = type === DEFAULT,
+        createdAt = createdAt!!
+    )
+}